// src/utils/storageResolver/dictionaryUtils/apiTypes/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  BigFloat: { input: any; output: any; }
  BigInt: { input: any; output: any; }
  Cursor: { input: any; output: any; }
  JSON: { input: any; output: any; }
};
// ...
export type Migration = {
  __typename?: 'Migration';
  id: Scalars['Int']['output'];
  name: Scalars['String']['output'];
  timestamp: Scalars['BigInt']['output'];
};
// ...
export type OmnipoolAssetDataConnection = {
  __typename?: 'OmnipoolAssetDataConnection';
  /** Aggregates across the matching connection (ignoring before/after/first/last/offset) */
  aggregates?: Maybe<OmnipoolAssetDatumAggregates>;
  /** A list of edges which contains the `OmnipoolAssetDatum` and cursor to aid in pagination. */
  edges: Array<OmnipoolAssetDataEdge>;
  /** Grouped aggregates across the matching connection (ignoring before/after/first/last/offset) */
  groupedAggregates?: Maybe<Array<OmnipoolAssetDatumAggregates>>;
  /** A list of `OmnipoolAssetDatum` objects. */
  nodes: Array<Maybe<OmnipoolAssetDatum>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** The count of *all* `OmnipoolAssetDatum` you could get from the connection. */
  totalCount: Scalars['Int']['output'];
};
// ...
export enum OmnipoolAssetDataOrderBy {
  AssetIdAsc = 'ASSET_ID_ASC',
  AssetIdDesc = 'ASSET_ID_DESC',
  AssetStateAsc = 'ASSET_STATE_ASC',
  AssetStateDesc = 'ASSET_STATE_DESC',
  BalancesAsc = 'BALANCES_ASC',
  BalancesDesc = 'BALANCES_DESC',
  IdAsc = 'ID_ASC',
  IdDesc = 'ID_DESC',
  Natural = 'NATURAL',
  ParaChainBlockHeightAsc = 'PARA_CHAIN_BLOCK_HEIGHT_ASC',
  ParaChainBlockHeightDesc = 'PARA_CHAIN_BLOCK_HEIGHT_DESC',
  PoolAddressAsc = 'POOL_ADDRESS_ASC',
  PoolAddressDesc = 'POOL_ADDRESS_DESC',
  PrimaryKeyAsc = 'PRIMARY_KEY_ASC',
  PrimaryKeyDesc = 'PRIMARY_KEY_DESC'
}

export type OmnipoolAssetDatum = {
  __typename?: 'OmnipoolAssetDatum';
  assetId: Scalars['Int']['output'];
  assetState: Scalars['JSON']['output'];
  balances: Scalars['JSON']['output'];
  id: Scalars['String']['output'];
  paraChainBlockHeight: Scalars['Int']['output'];
  poolAddress: Scalars['String']['output'];
};
// ...
export type OmnipoolAssetDatumFilter = {
  /** Checks for all expressions in this list. */
  and?: InputMaybe<Array<OmnipoolAssetDatumFilter>>;
  /** Filter by the object’s `assetId` field. */
  assetId?: InputMaybe<IntFilter>;
  /** Filter by the object’s `assetState` field. */
  assetState?: InputMaybe<JsonFilter>;
  /** Filter by the object’s `balances` field. */
  balances?: InputMaybe<JsonFilter>;
  /** Filter by the object’s `id` field. */
  id?: InputMaybe<StringFilter>;
  /** Negates the expression. */
  not?: InputMaybe<OmnipoolAssetDatumFilter>;
  /** Checks for any expressions in this list. */
  or?: InputMaybe<Array<OmnipoolAssetDatumFilter>>;
  /** Filter by the object’s `paraChainBlockHeight` field. */
  paraChainBlockHeight?: InputMaybe<IntFilter>;
  /** Filter by the object’s `poolAddress` field. */
  poolAddress?: InputMaybe<StringFilter>;
};
// ...
export type Query = {
  __typename?: 'Query';
  _squidStatus: Array<_ProcessorStatus>;
  migration?: Maybe<Migration>;
  /** Reads and enables pagination through a set of `Migration`. */
  migrations?: Maybe<MigrationsConnection>;
  /** Reads and enables pagination through a set of `OmnipoolAssetDatum`. */
  omnipoolAssetData?: Maybe<OmnipoolAssetDataConnection>;
  omnipoolAssetDatum?: Maybe<OmnipoolAssetDatum>;
  /**
   * Exposes the root query type nested one level down. This is helpful for Relay 1
   * which can only query top level fields if they are in a particular form.
   */
  query: Query;
  stablepool?: Maybe<Stablepool>;
  /** Reads and enables pagination through a set of `StablepoolAssetDatum`. */
  stablepoolAssetData?: Maybe<StablepoolAssetDataConnection>;
  stablepoolAssetDatum?: Maybe<StablepoolAssetDatum>;
  /** Reads and enables pagination through a set of `Stablepool`. */
  stablepools?: Maybe<StablepoolsConnection>;
  subProcessorStatus?: Maybe<SubProcessorStatus>;
  /** Reads and enables pagination through a set of `SubProcessorStatus`. */
  subProcessorStatuses?: Maybe<SubProcessorStatusesConnection>;
  xykPool?: Maybe<XykPool>;
  /** Reads and enables pagination through a set of `XykPoolAssetsDatum`. */
  xykPoolAssetsData?: Maybe<XykPoolAssetsDataConnection>;
  xykPoolAssetsDatum?: Maybe<XykPoolAssetsDatum>;
  /** Reads and enables pagination through a set of `XykPool`. */
  xykPools?: Maybe<XykPoolsConnection>;
};
// ...
export type Stablepool = {
  __typename?: 'Stablepool';
  fee: Scalars['Int']['output'];
  finalAmplification: Scalars['Int']['output'];
  finalBlock: Scalars['Int']['output'];
  id: Scalars['String']['output'];
  initialAmplification: Scalars['Int']['output'];
  initialBlock: Scalars['Int']['output'];
  paraChainBlockHeight: Scalars['Int']['output'];
  poolAddress: Scalars['String']['output'];
  poolId: Scalars['Int']['output'];
  /** Reads and enables pagination through a set of `StablepoolAssetDatum`. */
  stablepoolAssetDataByPoolId: StablepoolAssetDataConnection;
};
// ...
export type StablepoolAssetDatum = {
  __typename?: 'StablepoolAssetDatum';
  assetId: Scalars['Int']['output'];
  balances: Scalars['JSON']['output'];
  id: Scalars['String']['output'];
  paraChainBlockHeight: Scalars['Int']['output'];
  /** Reads a single `Stablepool` that is related to this `StablepoolAssetDatum`. */
  pool?: Maybe<Stablepool>;
  poolId?: Maybe<Scalars['String']['output']>;
};
// ...
export type StablepoolFilter = {
  /** Checks for all expressions in this list. */
  and?: InputMaybe<Array<StablepoolFilter>>;
  /** Filter by the object’s `fee` field. */
  fee?: InputMaybe<IntFilter>;
  /** Filter by the object’s `finalAmplification` field. */
  finalAmplification?: InputMaybe<IntFilter>;
  /** Filter by the object’s `finalBlock` field. */
  finalBlock?: InputMaybe<IntFilter>;
  /** Filter by the object’s `id` field. */
  id?: InputMaybe<StringFilter>;
  /** Filter by the object’s `initialAmplification` field. */
  initialAmplification?: InputMaybe<IntFilter>;
  /** Filter by the object’s `initialBlock` field. */
  initialBlock?: InputMaybe<IntFilter>;
  /** Negates the expression. */
  not?: InputMaybe<StablepoolFilter>;
  /** Checks for any expressions in this list. */
  or?: InputMaybe<Array<StablepoolFilter>>;
  /** Filter by the object’s `paraChainBlockHeight` field. */
  paraChainBlockHeight?: InputMaybe<IntFilter>;
  /** Filter by the object’s `poolAddress` field. */
  poolAddress?: InputMaybe<StringFilter>;
  /** Filter by the object’s `poolId` field. */
  poolId?: InputMaybe<IntFilter>;
};
// ...
export type StablepoolsConnection = {
  __typename?: 'StablepoolsConnection';
  /** Aggregates across the matching connection (ignoring before/after/first/last/offset) */
  aggregates?: Maybe<StablepoolAggregates>;
  /** A list of edges which contains the `Stablepool` and cursor to aid in pagination. */
  edges: Array<StablepoolsEdge>;
  /** Grouped aggregates across the matching connection (ignoring before/after/first/last/offset) */
  groupedAggregates?: Maybe<Array<StablepoolAggregates>>;
  /** A list of `Stablepool` objects. */
  nodes: Array<Maybe<Stablepool>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** The count of *all* `Stablepool` you could get from the connection. */
  totalCount: Scalars['Int']['output'];
};
// ...
export type XykPool = {
  __typename?: 'XykPool';
  assetAId: Scalars['Int']['output'];
  assetBId: Scalars['Int']['output'];
  id: Scalars['String']['output'];
  paraChainBlockHeight: Scalars['Int']['output'];
  poolAddress: Scalars['String']['output'];
  /** Reads and enables pagination through a set of `XykPoolAssetsDatum`. */
  xykPoolAssetsDataByPoolId: XykPoolAssetsDataConnection;
};
// ...
export type GetOmnipoolBlocksStorageStateQueryVariables = Exact<{
  filter?: InputMaybe<OmnipoolAssetDatumFilter>;
  first: Scalars['Int']['input'];
  offset: Scalars['Int']['input'];
  orderBy?: InputMaybe<Array<OmnipoolAssetDataOrderBy> | OmnipoolAssetDataOrderBy>;
}>;


export type GetOmnipoolBlocksStorageStateQuery = { __typename?: 'Query', omnipoolAssetData?: { __typename?: 'OmnipoolAssetDataConnection', totalCount: number, nodes: Array<{ __typename?: 'OmnipoolAssetDatum', assetId: number, assetState: any, balances: any, id: string, paraChainBlockHeight: number, poolAddress: string } | null> } | null };

export type GetXykPoolBlocksStorageStateQueryVariables = Exact<{
  filter?: InputMaybe<XykPoolFilter>;
  first: Scalars['Int']['input'];
  offset: Scalars['Int']['input'];
  orderBy?: InputMaybe<Array<XykPoolsOrderBy> | XykPoolsOrderBy>;
}>;


export type GetXykPoolBlocksStorageStateQuery = { __typename?: 'Query', xykPools?: { __typename?: 'XykPoolsConnection', totalCount: number, nodes: Array<{ __typename?: 'XykPool', assetAId: number, assetBId: number, id: string, paraChainBlockHeight: number, poolAddress: string, xykPoolAssetsDataByPoolId: { __typename?: 'XykPoolAssetsDataConnection', nodes: Array<{ __typename?: 'XykPoolAssetsDatum', assetId: number, balances: any, id: string, paraChainBlockHeight: number, poolId?: string | null } | null> } } | null> } | null };

export type GetStablepoolBlocksStorageStateQueryVariables = Exact<{
  filter?: InputMaybe<StablepoolFilter>;
  first: Scalars['Int']['input'];
  offset: Scalars['Int']['input'];
  orderBy?: InputMaybe<Array<StablepoolsOrderBy> | StablepoolsOrderBy>;
}>;


export type GetStablepoolBlocksStorageStateQuery = { __typename?: 'Query', stablepools?: { __typename?: 'StablepoolsConnection', totalCount: number, nodes: Array<{ __typename?: 'Stablepool', fee: number, finalAmplification: number, finalBlock: number, id: string, initialAmplification: number, initialBlock: number, paraChainBlockHeight: number, poolAddress: string, poolId: number, stablepoolAssetDataByPoolId: { __typename?: 'StablepoolAssetDataConnection', nodes: Array<{ __typename?: 'StablepoolAssetDatum', assetId: number, balances: any, id: string, paraChainBlockHeight: number, poolId?: string | null } | null> } } | null> } | null };


export const GetOmnipoolBlocksStorageState = gql`
    query GetOmnipoolBlocksStorageState($filter: OmnipoolAssetDatumFilter, $first: Int!, $offset: Int!, $orderBy: [OmnipoolAssetDataOrderBy!]) {
  omnipoolAssetData(
    filter: $filter
    orderBy: $orderBy
    first: $first
    offset: $offset
  ) {
    nodes {
      assetId
      assetState
      balances
      id
      paraChainBlockHeight
      poolAddress
    }
    totalCount
  }
}
    `;
export const GetXykPoolBlocksStorageState = gql`
    query GetXykPoolBlocksStorageState($filter: XykPoolFilter, $first: Int!, $offset: Int!, $orderBy: [XykPoolsOrderBy!]) {
  xykPools(filter: $filter, orderBy: $orderBy, first: $first, offset: $offset) {
    nodes {
      assetAId
      assetBId
      id
      paraChainBlockHeight
      poolAddress
      xykPoolAssetsDataByPoolId {
        nodes {
          assetId
          balances
          id
          paraChainBlockHeight
          poolId
        }
      }
    }
    totalCount
  }
}
    `;
export const GetStablepoolBlocksStorageState = gql`
    query GetStablepoolBlocksStorageState($filter: StablepoolFilter, $first: Int!, $offset: Int!, $orderBy: [StablepoolsOrderBy!]) {
  stablepools(filter: $filter, orderBy: $orderBy, first: $first, offset: $offset) {
    nodes {
      fee
      finalAmplification
      finalBlock
      id
      initialAmplification
      initialBlock
      paraChainBlockHeight
      poolAddress
      poolId
      stablepoolAssetDataByPoolId {
        nodes {
          assetId
          balances
          id
          paraChainBlockHeight
          poolId
        }
      }
    }
    totalCount
  }
}
    `;